Logging
This change adds Logging to the Cache download functionality. This required
the use of an extension to Logger to enable the typical first and subsequent
logging lines as well as standard printing of versions. The addition of this
Logger extension, also pointed out that we've got enough extension types that
extended Buildpack and Detect types would be useful. This change introduces
those as well.