Cache Avoidance
Previously when a file was downloaded into a cache it would be returned by
Artifact() but that caching wouldn't ever prevent additional downloads from
happening. This change adds cache avoidance to prevent future downloads from
occurring. This avoidance is built on top of a dependency.toml file being
written to the cache directory. If the dependency.toml file exists and
matches the dependency being downloaded (reflect.DeepEqual()), the download
is avoided and the path returned immediately. If the file does not exist, or
does not match, then the dependency is downloaded.