Dependency Cache
Previously when a dependency was contributed into a cache that caching
wouldn't ever prevent addition contributions from happening.  This change adds
contribution avoidance to prevent future contributions from occurring.  This
avoidance is built on top of a dependency.toml file being written to the cache
directory.  If the dependency.toml file exists and matches the dependency
being contributed (reflect.DeepEqual()), the contribution is avoided and the
the function returns immediately.  If the file does not exist, or does not
match, the dependency is contributed.