Dependencies Detected in Buildpack Cache
Previously if dependencies were cached within a buildpack (for an air-gapped
installation) these dependencies were ignored when it came time to resolve a
dependency.  This change updates the DownloadCacheLayer to determine if a
dependency has been cached within the buildpack and if so, return that
artifact path.  If it hasn't been cached, the layer will check the cache for a
previous download.  If it isn't found there, it will finally be downloaded
from the specified URI.

[resolves #4]