Create Layer Directories
Previously, when a layer was contributed to, each contribution callback was
responsible for creating any necessary parent directories in order to write to
the layer.  While the library was well-written enough to ensure that this
always happened, it was very easy to write code that did not do this and would
cause issues.  This change updates the the Contribute() functions such that
they always create the necessary directories before the callback is called.

Signed-off-by: Ben Hale <bhale@pivotal.io>