Layer Improvements
Previously, there was a bit of superfluous logging when using a download layer
due to the way it leveraged the existing Layer.Contribute functionality.  In
order to avoid this logging, and to make the behavior generally more
efficient, this change extracts metadata matching logic into code that can be
called directly, and not be hidden in Layer.Contribute.

In addition, this change updates the tests to be randomized.

Signed-off-by: Ben Hale <bhale@pivotal.io>