Tolerate Invalid Layer Metadata
Previously if invalid layer metadata was encountered, e.g. if the structure of
the layer metadata had changed, and error was returned.  This error would
interrupt all execution of layer contribution.  This change updates that
behavior to indicate that the layer metadata does not match rather than being
an error which allows contribution to proceed and overwrite the older
metadata.

Signed-off-by: Ben Hale <bhale@pivotal.io>