Changing Package Location
Previously the artifact that was created as part of packaging went to a name
and location that was Artifactory/Maven friendly.  This change updates the
archive to have a simple, obvious name for and location for the artifact.  The
complexity of making the artifact Aritfactory/Maven friendly then falls to the
buildpacks.

Signed-off-by: Ben Hale <bhale@pivotal.io>