Automated Layer Cleanup
Previously when a buildpack ran if some bit of its implementation or
configuration that caused the creation of layers had changed, those layers
would not be cleaned up.  Only buildpack authors could do this and manual
management meant that most forgot.  This change now keeps track of all the
layers that a buildpack touches during a run (including any transitive
download layers) and compares them with the list of layers set to be added to
the image.  Every extraneous image is then removed while logging that it
happened.

Signed-off-by: Ben Hale <bhale@pivotal.io>