Contributed Dependencies in Build Plan
Previously, if a buildpack wanted to signal that it had consumed a dependency
in its buildplan it had to be done manually.  This change updates the library
so that any DependencyLayer that is contributed (even if avoided) contributes
to the BuildPlan signaling consumption.  The buildpack can still signal
consumption through Build.Success() and any overlapping keys use the
explicitly declared values from the buildpack.

Signed-off-by: Ben Hale <bhale@pivotal.io>