Manual Layer Touch
Previously, the only way a layer could be marked as used by a build would be
perform some operation such as Contribute() or WriteProfile() and have it
automatically marked.  However, there are layers that are used simply as
caches that never have these functions called and therefore are always
removed.  This change adds a new Layer.Touch() function that allows it to be
manually marked for retention.

Signed-off-by: Ben Hale <bhale@pivotal.io>