Do Not Delete Invalid Layers
Previously when a layer was judged to be invalid (i.e. metadata did not match
expected), it was removed and then the contributor was called.  This was
generally a good policy as it ensured that no remnants were left when the
contribution occurred.  However, there are times (e.g. npm install) when using
the previous contents of a layer is desirable.

To try and balance these two competing requirements the removal of the
directory now happens above the base Layer type and is automatic for
DownloadLayers and DependencyLayers.  This means that the most common type of
contribution (a layer that explodes a dependency) will continue to behave in a
way that minimizes the amount of code a user needs to write.  If you
contribute directly using the base Layer, then you are responsible for
deciding whether or not to clean the directory.

[resolves #6]

Signed-off-by: Ben Hale <bhale@pivotal.io>