Services
This change adds support for services to the library.  It adds functionality
for determining if a dependency is bound while the buildpack is running which
is used to determine whether to contribute a dependency.  It also adds a
helper method (to be used at launch time !!) to extract the values of service
credentials.

Signed-off-by: Ben Hale <bhale@pivotal.io>