Identify Buildpack In User Agent
Previously, when the buildpack would download a dependency, it would identify
itself as a generic Go application.  Given some recent analysis of dependency
downloads, it is useful to have some insight into what buildpacks are being
used to download dependencies.  This change updates the User-Agent header of
those downloads to identify the buildpack and version.

Signed-off-by: Ben Hale <bhale@pivotal.io>