Buildpack exposes default_versions properties in Metadata
Previously, given a dependency with multiple versions in a buildpack, there
was no consistent way to define the default version.  This change adds support
for a default_versions property hash in the buildpack.toml file that exposes a
map from a dependency id to a default version (represented by a semver
constraint).  In addition, support was added to the TestBuildFactory for
configuring this value, and a new test matcher was added to ensure that the
proper version was contributed.

[#163528537][#11]

Co-authored-by: Guillermo Kardolus <gkardolus@pivotal.io>
Co-authored-by: Don Nguyen <dnguyen@pivotal.io>
Co-authored-by: David Freilich <dfreilich@pivotal.io>
Signed-off-by: Ben Hale <bhale@pivotal.io>