Promote Runner Abstraction
Previously the runner abstraction was a helper type that implemented an
interface that was never defined in the library.  This change promotes the
runner to a top-level type, creates an interface, production implementation,
and test implementation.  It also adds it to the Detect, Build,
TestDetectFactory, and TestBuildFactory to facility swapping the production
implementation with the test implementation.

Signed-off-by: Ben Hale <bhale@pivotal.io>