Improved Non-Match Message
Previously, when the metadata did not match, an error saying it did not match
was printed.  This wasn't as useful as it might have been because it did not
describe the discrepancy.  This change updates the code to print the expected
and actual metadata as well in order to facilitate comparison.

[resolves #19][#164601296]

Co-authored-by: Don Nguyen <dnguyen@pivotal.io>
Signed-off-by: Ben Hale <bhale@pivotal.io>