Improve Metadata Matcher Message
Previously, when the layer or persistent metadata did not match the expectation, an error saying it did not match was printed. This wasn't as useful as it might have been, because it did not describe the discrepency. This change updates the code to print the expected and actual metadata, in order to facilitate comparison.