MultiDependencyLayer
This change adds a MultiDependencyLayer.  This layer works much like the
DependencyLayer, except it takes a collection of dependencies, and a map of
dependency-id to contributor functions.  In this way, multiple dependencies
can be contributed to the same layer, and avoidance can be only be done when
all dependencies are up to date.

Signed-off-by: Ben Hale <bhale@pivotal.io>