Logging Improvements
This change constitutes an overhaul of the logging system and output format.
These changes introduce a number of new functions and deprecate most of the
previous ones.

The new structure is centered around a Title, Header, and Body.  Each of these
has a limited indention (0, 2, 4 spaces) and the body is styled as faint to
ensure that it fades to the background, leaving the header as the typically
"lowest" abstraction that matters.