Remove Environment SetAll
Previously, the lifecycle did not load the entire environment when a buildpack
was invoked.  Because CF buildpacks never used the environmental
differentiation that filesystem-based environment variables provided, this
library always loaded them by default.  With a recent change to the spec and
lifecycle, all environment variables are automatically loaded before invoking
the buildpack (with explicit exceptions) and therefore this behavior is no
longer required.

Signed-off-by: Ben Hale <bhale@pivotal.io>