Terminal Error
Previously when a buildpack failed, especially if there was a lot of output,
it wasn't clear which buildpack had just failed (finding the announcement was
difficult).  This change adds a TerminalError function to logger to
re-announce the buildpack that has just failed and then print the error
message.  This should be used in the boilerplate stanza in the main()
functions that used to print the error message before os.Exit()-ing.

Signed-off-by: Ben Hale <bhale@pivotal.io>