Do Not Cache Downloads
Previously, when a dependency was downloaded it was placed in a layer that was
marked for build caching.  It was pointed out that there was no real way for
this cached download to ever be used again.  Given that fact (which took some
real soul searching to admit), and the performance penalty of calculating
hashes and moving data around, coupled with the most unforgivable sin, log
output that didn't look good, downloads should not be cached at all after they
are used.

Special thanks to @ekcasey for keeping on me until I saw what she already
knew.

Signed-off-by: Ben Hale <bhale@pivotal.io>