Runner Detects TTY
Previously, the CommandRunner would execute and specified command in an
environment that was not a TTY, thus causing most commands to disable their
optional coloring via escape codes.  This wasn't great because a lot of tools
use color very well to help emphasize or de-emphasize their logging.

This change adds support for both detecting when the buildpack is running in a
TTY and then passing that information on to the tools that are being executed.
This results in non-colored text if the buildpack is in a non-colored
environment and colored text if the buildpack is in a colored environment.

Note that this will always be disabled in pack at this moment because pack
does not properly communicate the same information.  A change will be coming
in pack to properly execute buildpacks in the same TTY state as pack runs in.

Signed-off-by: Ben Hale <bhale@pivotal.io>